Extract menu code from main code
First pass at removing the generic menu code from the device-specific
main code. There's clearly still lots of room for improvement, but it's
still functional and headed in the right way, imo.
