+++ code/veccart/main.c
@@ -31,13 +31,13 @@
 #include "led.h"
 #include "delay.h"
 #include "main.h"
+#include "menu.h"
 #include "msc.h"
 
 //#include "rom.h"
 #include "xprintf.h"
 #include "fatfs/ff.h"
 
-#define MENU_TEXT_LEN 20
 
 //Memory for the menu ROM and the running cartridge.
 //We keep both in memory so we can quickly exchange them when a reset has been detected.
@@ -47,16 +47,6 @@ char *romData=menuData;
 unsigned char parmRam[256];
 
 char menuDir[_MAX_LFN + 1];
-
-typedef struct {
-	int is_dir;
-	char fname[_MAX_LFN + 1];
-} file_entry;
-
-typedef struct {
-	int num_files;
-	file_entry f_entry[80];
-} dir_listing;
 
 union cart_and_listing {
 	char cartData[64*1024];
@@ -114,137 +104,6 @@ void loadRom(char *fn) {
 	f_close(&f);
 }
 
-int removeExtension(char* filename, char* extension) {
-	char* ptr = strstr(filename,extension);
-	if (ptr) {
-		*ptr = 0;
-		return 1;
-	}
-	return 0;
-}
-
-static int f_entry_compare(const void* a, const void* b)
-{
-	// setting up rules for comparison
-	file_entry *f_entryA = (file_entry *)a;
-	file_entry *f_entryB = (file_entry *)b;
-
-	int dirA = f_entryA->is_dir;
-	int dirB = f_entryB->is_dir;
-
-	if (dirA == dirB) {
-		return (strcmp(f_entryA->fname, f_entryB->fname));
-	} else {
-		 return dirB - dirA;
-	}
-}
-
-void sortDirectory(char *fdir) {
-	DIR d;
-	FILINFO fi;
-	char lfn[_MAX_LFN + 1];
-	fi.lfname=lfn;
-	fi.lfsize=sizeof(lfn);
-
-	int idx = 0;
-	dir_listing *d_listing = &(c_and_l.listing);
-	file_entry *f_entry;
-	char *name;
-	int is_dir;
-
-	// initial unsorted directory read
-	f_opendir(&d, fdir);
-	while (f_readdir(&d, &fi)==FR_OK) {
-		// xprintf("Found file %s (%s)\n", fi.lfname, fi.fname);
-		f_entry=&(d_listing->f_entry[idx]);
-
-		if (fi.fname[0]==0) break;
-		if (fi.fname[0]=='.') {
-			if (fi.fname[1] != '.')
-				continue;
-			if (strcmp(fdir, "/roms") == 0)
-				continue;
-		}
-
-		is_dir = (fi.fattrib & AM_DIR) ? 1 : 0;
-		if (fi.lfname[0]=='.') continue; // ignore MacOS dotfiles
-		name=fi.lfname;
-		if (name==NULL || name[0]==0) name=fi.fname; //use short name if no long name available
-
-		f_entry->is_dir = is_dir;
-		strcpy(f_entry->fname, name);
-		idx++;
-	}
-	f_closedir(&d);
-
-	d_listing->num_files = idx;
-
-	xprintf("Found %d files in %s\n", d_listing->num_files, fdir);
-
-	// xprintf("About to sort\n");
-	qsort(d_listing->f_entry, d_listing->num_files, sizeof(file_entry), f_entry_compare);
-	// xprintf("Done sorting\n");
-}
-
-//Get a listing of the roms in the 'roms/' directory and
-//poke them into the menu cartridge space
-void loadListing(char *fdir) {
-	char buff[30];
-	char *name;
-
-	int ptrpos = menuIndex + 1;  // fixed location in multicart.bin for 512 filename pointers (from &ptrpos ~ &strpos)
-	int strpos = ptrpos + 0x200; // filename data starts here
-
-	int i;
-	int is_dir;
-	int idx;
-	file_entry *f_entry;
-
-	xprintf("Reading dir: %s...\n", fdir);
-	sortDirectory(fdir);
-	for (idx = 0; idx < c_and_l.listing.num_files; idx++) {
-		f_entry=&(c_and_l.listing.f_entry[idx]);
-
-		strncpy(buff, f_entry->fname, sizeof(buff));
-		name=buff;
-
-		is_dir=f_entry->is_dir;
-
-		//xprintf("Found file %s (%d), file %d\n", name, is_dir, idx);
-		romData[ptrpos++]=strpos>>8;
-		romData[ptrpos++]=strpos&0xff;
-
-		// remove extensions
-		removeExtension(name, ".bin");
-		removeExtension(name, ".BIN");
-		removeExtension(name, ".vec");
-		removeExtension(name, ".VEC");
-
-		i = is_dir ? (MENU_TEXT_LEN-2) : MENU_TEXT_LEN;
-		if (is_dir) romData[strpos++]='<';
-		while (*name!=0 && i>0) {
-			if (*name<32) {
-				romData[strpos++]=' ';
-			} else if (*name>=32 && *name<95) {
-				romData[strpos++]=*name;
-			} else if (*name>='a' && *name<='z') {
-				romData[strpos++]=(*name-'a')+'A'; //convert to caps
-			} else {
-				romData[strpos++]='_';
-			}
-			name++;
-			i--;
-		}
-		if (is_dir) romData[strpos++]='>';
-		romData[strpos++]=0x80; //end of string
-	}
-	//finish with zero ptr
-	romData[ptrpos++]=0;
-	romData[ptrpos++]=0;
-	xprintf("Done.\n");
-}
-
-
 //Stream data, for the Bad Apple demo
 //Name of the file sucks and I'd like to make a rpc function that's a bit
 //more universal... you should eg be able to pass through the name of the
@@ -269,7 +128,7 @@ void doChangeRom(char* basedir, int i) {
 
 	menuData[menuIndex]=i; //save selection so we can go back there after reset
 	xprintf("Changing to rom no %d in %s\n", i, basedir);
-	sortDirectory(basedir); // recreate file listing, as loading a cart overwrote the union
+	sortDirectory(basedir, &c_and_l.listing); // recreate file listing, as loading a cart overwrote the union
 	file_entry f = c_and_l.listing.f_entry[i];
 
 	if (f.is_dir) {
@@ -284,7 +143,7 @@ void doChangeRom(char* basedir, int i) {
 		}
 
 		romData=menuData;
-		loadListing(menuDir);
+		loadListing(menuDir, &c_and_l.listing, menuIndex+1 , menuIndex+1+0x200, romData);
 		menuData[menuIndex]=0; //save selection so we can go back there after reset
 
 		xprintf("Done listing for : %s\n", menuDir);
@@ -441,7 +300,7 @@ int main(void) {
 		strncpy(menuDir, "/roms", sizeof(menuDir));
 		f_mount(&FatFs, "", 0);
 		loadRom("/multicart.bin");
-		loadListing(menuDir);
+		loadListing(menuDir, &c_and_l.listing, menuIndex+1 , menuIndex+1+0x200, romData);
 
 		//Go emulate a ROM.
 		SYSCFG_MEMRMP=0x3; //mak ram at 0
